+++ backend/diet-generator/src/main/java/org/ectimel/dietgenerator/domain/model/Proteins.java
@@ -3,8 +3,10 @@ package org.ectimel.dietgenerator.domain.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class Proteins {
-    private Double totalProteins;
+    private BigDecimal totalProteins;
 }
